% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/put-and-pull.R
\name{hb.pull}
\alias{hb.pull}
\title{Pull Data from HBase in tidyr principles}
\usage{
hb.pull(tablename, column_family, start = "", end = NULL, timestamp = 0,
  caching = 0, filterstring = character(0), hbc = hb.defaults("hbc"),
  sz = function(x) serialize(x, connection = NULL), usz = unserialize, ...)
}
\arguments{
\item{tablename}{The name of the desired HBase table}

\item{column_family}{The column family desired - see details for specifying column}

\item{start}{The starting point a scan - first desired rowkey}

\item{end}{The end poit of a HBase scan - last desired rowkey}
}
\description{
Retrieve data from HBase in a tidry data.frame format.
}
\details{
This function serves to make pulling from HBase in R much more straightforward, particularily on how you receive the
data. Following the pricipals that tidyr has set to accomplish, the retrieved valued is a data.frame in the following format:
}
\author{
Aaron Benz
}

